+++ pkg/operator/ceph/cluster/osd/provision_spec.go
@@ -158,7 +158,7 @@ func (c *Cluster) provisionPodTemplateSpec(osdProps osdProperties, restart v1.Re
 	} else {
 		osdProps.getPreparePlacement().ApplyToPodSpec(&podSpec)
 	}
-	removeDuplicateEnvVars(&podSpec)
+	k8sutil.RemoveDuplicateEnvVars(&podSpec)
 
 	podMeta := metav1.ObjectMeta{
 		Name: AppName,
